requirements.txt
altair==5.5.0
anyio==4.8.0
asgiref==3.8.1
asttokens==3.0.0
attrs==24.3.0
backports.tarfile==1.2.0
bcrypt==4.2.1
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
debugpy==1.8.9
decorator==5.1.1
Django==5.1.6
django-appconf==1.1.0
django-imagekit==5.0.0
django-multiselectfield==0.1.13
django-widget-tweaks==1.5.0
et_xmlfile==2.0.0
executing==2.1.0
extra-streamlit-components==0.1.71
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
fonttools==4.56.0
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
ipykernel==6.29.5
ipython==8.30.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
keyring==25.6.0
kiwisolver==1.4.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
more-itertools==10.6.0
msal==1.31.1
mysql==0.0.3
mysql-connector-python==9.1.0
mysqlclient==2.2.6
narwhals==1.21.1
nest-asyncio==1.6.0
numpy==2.2.1
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pilkit==3.0
pillow==11.1.0
platformdirs==4.3.6
plotly==5.24.1
prompt_toolkit==3.0.48
protobuf==5.29.2
psutil==6.1.0
pure_eval==0.2.3
pyarrow==18.1.0
pycparser==2.22
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.10.1
PyMySQL==1.1.1
pyparsing==3.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pywin32==308
pywin32-ctypes==0.2.3
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
selenium==4.29.0
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.36
sqlparse==0.5.3
stack-data==0.6.3
streamlit==1.38.0
streamlit-authenticator==0.3.3
tenacity==8.5.0
threadpoolctl==3.5.0
toml==0.10.2
tornado==6.4.2
traitlets==5.14.3
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
watchdog==4.0.2
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
WTForms==3.2.1
zipp==3.21.0
